backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for the multi-user todo web application with JWT authentication"
authors = ["Developer <developer@example.com>"]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sqlmodel>=0.0.16",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "better-exceptions>=0.3.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0"
]